Solved 16953 / Silver II

+++ 백준/분류별/greedy/BOJ_16953.java
@@ -12,8 +12,8 @@ public class BOJ_16953 {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
 
-        long a = Integer.parseInt(st.nextToken());
-        long b = Integer.parseInt(st.nextToken());
+        long a = Long.parseLong(st.nextToken());
+        long b = Long.parseLong(st.nextToken());
 
         int cnt = 1;
         while (b != a) {
